uprade to react 18 and react-router-dom v6, discard plugins
