+++ client/src/pages/search/Styles.ts
