+++ client/src/pages/system/BaseCard.tsx
